requirements.txt
python-dotenv>=1.0.0
langchain>=0.0.300
langchain_openai
openai>=0.27.0
pytest>=7.0.0
rich>=13.0.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
sse-starlette>=1.0.7

